Redo data model for area designer to be more resilient against data loss.
Previously, any error which resulted in a blank database could write
that entire state back and erase all data. In this model, the web UI can
never change the list of areas, only the points within. So it's still
possible to clobber work scoped to an area, but that's a tradeoff that
I'm ok with given the amount of work that goes into anyone area is
likely to be small, and it's much harder clobber it.